+++ typeit/__init__.py
@@ -1,5 +1,6 @@
 from . import flags
-from .parser import type_constructor
+from . import sums
+from .combinator import type_constructor
 from .schema.errors import Error
 
-__all__ = ('type_constructor', 'flags', 'Error')
+__all__ = ('type_constructor', 'flags', 'sums', 'Error')