+++ CHANGELOG.rst
@@ -2,6 +2,14 @@
 CHANGELOG
 =========
 
+0.15.0
+==============
+
+* SumType from/to mapping #43
+* Support explicit tags for SumType Variants #32
+* Parser memoization #38
+
+
 0.14.1
 ==============